task_service/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pika
aio_pika
python-dotenv
python-jose[cryptography]